Moved replace_parameters() & associated function into _utils/param_validation.py
 - Tweaked docstring & symbol names for replace_parameters decorator
 - Might create a _utils/misc.py and move it there?
