+++ nilearn/plotting/html_connectome.py
@@ -4,6 +4,8 @@ import warnings
 
 import numpy as np
 from scipy import sparse
+
+from nilearn._utils.param_validation import replace_parameters
 from .. import datasets
 from . import cm
 
@@ -85,41 +87,6 @@ def _make_connectome_html(connectome_info, embed_js=True):
     return ConnectomeView(as_html)
 
 
-def replace_parameters(replacement_params, end_version, lib_name):
-    """
-    Decorator to deprecate & replace specificied parameters
-    in the decorated functions and methods.
-    
-    Add **kwargs as the last parameter in the decorated method/function.
-    
-    Parameters
-    ----------
-    replacement_params : Dict[string, string]
-        Dict where the key-value pairs represent the old parameters
-        and their corresponding new parameters.
-        Example: {old_param1: new_param1, old_param2: new_param2,...}
-        
-    end_version : str
-        Version when the deprecated parameters will cease functioning
-        and no more wrnings will be displayed.
-        Example: '0.6.0'
-        
-    lib_name: str
-        Name of the library to which the decoratee belongs.
-        Example: Nilearn, Nistats
-    """
-    
-    def _replace_params(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            _warn_deprecated_params_view_connectome(replacement_params, end_version, lib_name, kwargs)
-            kwargs = _transfer_deprecated_param_vals(replacement_params, kwargs)
-            return func(*args, **kwargs)
-        
-        return wrapper
-    return _replace_params
-
-
 def _replacement_params_view_connectome():
     """ Returns a dict containing deprecated & replacement parameters
         as key-value pair.
@@ -194,39 +161,6 @@ def view_connectome(adjacency_matrix, node_coords, edge_threshold=None,
         adjacency_matrix, node_coords, threshold=edge_threshold, cmap=edge_cmap,
         symmetric_cmap=symmetric_cmap, marker_size=node_size)
     return _make_connectome_html(connectome_info)
-
-
-def _warn_deprecated_params_view_connectome(replacement_params, end_version, lib_name, kwargs):
-    """ For the decorator replace_parameters(),
-        raises warnings about deprecated parameters.
-    """
-    used_deprecated_params = set(kwargs).intersection(replacement_params)
-    for deprecated_param_ in used_deprecated_params:
-        replacement_param = replacement_params[deprecated_param_]
-        param_deprecation_msg = (
-            'The parameter "{}" will be removed in {} version {}. '
-            'Please use the parameter "{}" instead.'.format(deprecated_param_,
-                                                            lib_name,
-                                                            end_version,
-                                                            replacement_param,
-                                                            )
-        )
-        warnings.filterwarnings('always', message=param_deprecation_msg)
-        warnings.warn(category=DeprecationWarning,
-                      message=param_deprecation_msg,
-                      stacklevel=3)
-
-
-def _transfer_deprecated_param_vals(replacement_params, kwargs):
-    """ For the decorator replace_parameters(), reassigns new parameters
-    the values passed to their corresponding deprecated parameters.
-    """
-    for old_param, new_param in replacement_params.items():
-        old_param_val = kwargs.setdefault(old_param, None)
-        if old_param_val is not None:
-            kwargs[new_param] = old_param_val
-        kwargs.pop(old_param)
-    return kwargs
 
 
 def _deprecate_params_view_markers(func):
